+++ quackosm/_rich_progress.py
@@ -5,7 +5,8 @@ from collections.abc import Iterable
 
 __all__ = ["TaskProgressSpinner", "TaskProgressBar"]
 
-TOTAL_STEPS = 36
+TOTAL_STEPS = 33
+
 
 class TaskProgressSpinner:
     def __init__(self, step_name: str, step_number: str):
@@ -51,11 +52,23 @@ class TaskProgressBar:
                 BarColumn,
                 MofNCompleteColumn,
                 Progress,
+                ProgressColumn,
                 SpinnerColumn,
+                Task,
+                Text,
                 TextColumn,
                 TimeElapsedColumn,
                 TimeRemainingColumn,
             )
+
+            class SpeedColumn(ProgressColumn):
+                def render(self, task: "Task") -> Text:
+                    if task.speed is None:
+                        return Text("")
+                    elif task.speed >= 1:
+                        return Text(f"{task.speed:.2f} it/s")
+                    else:
+                        return Text(f"{1/task.speed:.2f} s/it") # noqa: FURB126
 
             self.progress = Progress(
                 SpinnerColumn(),
@@ -68,8 +81,10 @@ class TaskProgressBar:
                 MofNCompleteColumn(),
                 TextColumn("•"),
                 TimeElapsedColumn(),
-                TextColumn("•"),
+                TextColumn("<"),
                 TimeRemainingColumn(),
+                TextColumn("•"),
+                SpeedColumn(),
                 transient=False,
                 speed_estimate_period=1800,
             )